+++ CHANGELOG.md
@@ -31,6 +31,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Remove `CourseProductItem` as a widget
 
+### Changed
+
+- In `SaleTunnelStepValidation`, allow user to show/hide target course details
+
 ## [2.23.0] - 2023-07-28
 
 ### Added